Crate threads

AlGlobo - Simple HTTP server for reserving a flight

This server receives POST calls from the client and attempts to make a reservation to the corresponding airline

Start the server with cargo run or even cargo run airlines.txt to specify the airlines file (specified on the airlines.rs module)

The requests are simple POSTs to the route / with a JSON body formatted as

 {
  "origin": "EZE", // Origin airport
  "destination": "JFK", // Destination airport
  "airline": "AA", // Airline code, which of course should be present on our config file
  "hotel": true // If the client wants to send a package, it must go through the hotel server
 }

This example could be sent with curl as: curl -i -d '{"origin":"EZE", "destination":"JFK", "airline":"AA", "hotel":true}' -H "Content-Type: application/json" -X POST http://localhost:8080/

Response Status of the actix-web server:

200 if the request was successful -> this doesn’t mean that the flight reservation was successful! It just means that the communication was successful

406 if the JSON was valid, but the airline is not supported by our server (not present on the config file)

400 if the request JSON is not valid

The server has a thread always listening to keyboard events. If the user presses s the server will show the flight stats, and if the user presses q the server will gracefully exit.

Modules

airlines

Handle airlines config

alglobo

Make the reservations

keyboard

statistics

Flight Stats

Structs

AppState

This is the shared state that will be shared across every thread listening to new requests: the airlines configurations and the universal stats entity

reservation

Functions

main

The main function. It starts a thread for the keyboard listener, and it starts the actix-web server